+++ gui/rescodes.hpp
@@ -1,6 +1,24 @@
 #ifndef _APP_RESOURCE_CODES
 #define _APP_RESOURCE_CODES
 
+	//	gui
+	#define GUI_LOGWIN				11
+	#define GUI_COMPROM				12
+	#define GUI_COMBO_PORT			13
+	#define GUI_COMBO_SPEED			14
+	#define GUI_BTN_SEND			15
+	#define GUI_BTN_CLR				16
+	#define GUI_BTN_UPD				17
+	#define GUI_CHK_NLN				18
+	#define GUI_CHK_QKAT			19
+
+	#define GUI_AT_PREF				20
+	#define GUI_AT_AT				21
+	#define GUI_AT_OK				22
+	#define GUI_AT_ID				23
+
+
+	//	other
 	#define CM_FILE_SVLOG			9070
 	#define CM_FILE_EXIT			9071
 	#define CM_ABOUT				9072
